+++ tests/qgis/processings/conftest.py
@@ -54,3 +54,21 @@ def photon_srv(monkeypatch) -> pytest_httpserver.HTTPServer:
     server.clear()
     if server.is_running():
         server.stop()
+
+
+@pytest.fixture(scope="function")
+def rnb_srv(monkeypatch) -> pytest_httpserver.HTTPServer:
+    """Fixture to start simulating rnb-api.beta.gouv.fr server and monkeypatch plugin settings for url
+
+    :param monkeypatch: monkeypatch for PlgSettingsStructure
+    :type monkeypatch: MonkeyPatch
+    :yield: server started
+    :rtype: pytest_httpserver.HTTPServer
+    """
+    server = pytest_httpserver.HTTPServer()
+    server.start()
+    monkeypatch.setattr(PlgSettingsStructure, "rnb_url", server.url_for(""))
+    yield server
+    server.clear()
+    if server.is_running():
+        server.stop()